Move fswatcher to internal/
