+++ internal/fswatcher/fswatcher_test.go
